Task Completion, Progress Bar, Randomly-Assigned Static Tasks
Added functionality where every user will have random tasks generated. Added functionality to remove the HTML element displaying the associated task. Also added a progress bar that can be updated based off of JavaScript variables.

+++ functions.php
@@ -11,9 +11,11 @@ if ($connection->connect_error)
     die("Fatal Error 1");
 
 function createTable($name, $query){
+  if ($name != 'tasks'){
     queryMysql("DROP TABLE $name");
-    queryMysql("CREATE TABLE IF NOT EXISTS $name($query)");
-    echo "Table '$name' created or already exists.<br>";
+  }
+  queryMysql("CREATE TABLE IF NOT EXISTS $name($query)");
+  echo "Table '$name' created or already exists.<br>";
 }
 
 function getRows(){
